# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Corporate Gifting ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: industry, in: query, schema: { type: string } }
      responses:
        "200":
          description: Clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /v1/clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
  /v1/clients/{id}/interactions:
    get:
      operationId: listClientInteractions
      tags: [clients]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Interactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientInteraction"
    post:
      operationId: createClientInteraction
      tags: [clients]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInteractionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientInteraction"

  # ── Products ──────────────────────────────────────────────────────────────
...
  /v1/inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      parameters:
        - { name: productId, in: query, schema: { type: integer } }
        - { name: lowStock, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Inventory records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /v1/inventory/movements:
    get:
      operationId: listInventoryMovements
      tags: [inventory]
      parameters:
        - { name: productId, in: query, schema: { type: integer } }
        - { name: type, in: query, schema: { type: string } }
      responses:
        "200":
          description: Movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryMovement"
    post:
      operationId: createInventoryMovement
      tags: [inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryMovementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryMovement"

  # ── Assembly Jobs ─────────────────────────────────────────────────────────
...
  /v1/invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: clientId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /v1/invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # ── Payments ──────────────────────────────────────────────────────────────
  /v1/payments:
    get:
      operationId: listPayments
      tags: [payments]
      parameters:
        - { name: invoiceId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── Analytics ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, companyName, contactPerson, email, createdAt]
      properties:
        id: { type: integer }
        companyName: { type: string }
        contactPerson: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        gstNumber: { type: ["string", "null"] }
        industry: { type: ["string", "null"] }
        tags: { type: ["string", "null"] }
        billingAddress: { type: ["string", "null"] }
        shippingAddress: { type: ["string", "null"] }
        createdAt: { type: string }
    ClientInput:
      type: object
      required: [companyName, contactPerson, email]
      properties:
        companyName: { type: string }
        contactPerson: { type: string }
        email: { type: string }
        phone: { type: string }
        gstNumber: { type: string }
        industry: { type: string }
        tags: { type: string }
        billingAddress: { type: string }
        shippingAddress: { type: string }
    ClientUpdate:
      type: object
      properties:
        companyName: { type: string }
        contactPerson: { type: string }
        email: { type: string }
        phone: { type: string }
        gstNumber: { type: string }
        industry: { type: string }
        tags: { type: string }
        billingAddress: { type: string }
        shippingAddress: { type: string }
    ClientInteraction:
      type: object
      required: [id, clientId, type, notes, createdAt]
      properties:
        id: { type: integer }
        clientId: { type: integer }
        type: { type: string }
        notes: { type: string }
        createdAt: { type: string }
    ClientInteractionInput:
      type: object
      required: [type, notes]
      properties:
        type: { type: string }
        notes: { type: string }

    # ── Products ──
...
    InventoryItem:
      type: object
      required: [productId, productName, stockLevel, lowStockThreshold]
      properties:
        productId: { type: integer }
        productName: { type: string }
        category: { type: string }
        stockLevel: { type: integer }
        lowStockThreshold: { type: integer }
        isLowStock: { type: boolean }
    InventoryMovement:
      type: object
      required: [id, productId, productName, type, quantity, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        type: { type: string }
        quantity: { type: integer }
        batch: { type: ["string", "null"] }
        reference: { type: ["string", "null"] }
        createdAt: { type: string }
    InventoryMovementInput:
      type: object
      required: [productId, type, quantity]
      properties:
        productId: { type: integer }
        type: { type: string }
        quantity: { type: integer }
        batch: { type: string }
        reference: { type: string }

    # ── Assembly ──
...
    Invoice:
      type: object
      required: [id, invoiceNumber, salesOrderId, clientId, clientName, totalAmount, gstAmount, grandTotal, status, createdAt]
      properties:
        id: { type: integer }
        invoiceNumber: { type: string }
        salesOrderId: { type: integer }
        orderNumber: { type: ["string", "null"] }
        clientId: { type: integer }
        clientName: { type: string }
        totalAmount: { type: number }
        gstAmount: { type: number }
        grandTotal: { type: number }
        status: { type: string }
        dueDate: { type: ["string", "null"] }
        createdAt: { type: string }
    InvoiceInput:
      type: object
      required: [salesOrderId, gstPercent]
      properties:
        salesOrderId: { type: integer }
        gstPercent: { type: number }
        dueDate: { type: string }

    # ── Payments ──
    Payment:
      type: object
      required: [id, invoiceId, amount, type, paymentDate, createdAt]
      properties:
        id: { type: integer }
        invoiceId: { type: integer }
        amount: { type: number }
        type: { type: string }
        paymentDate: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    PaymentInput:
      type: object
      required: [invoiceId, amount, type, paymentDate]
      properties:
        invoiceId: { type: integer }
        amount: { type: number }
        type: { type: string }
        paymentDate: { type: string }
        notes: { type: string }

    # ── Shared ──
...